Make interpretation of local rotation matrices consistent
Previously, the ROTLOC matrices were read in transpose in
struct_read().  They were then used as is in ‘w2w’ but transposed (to
yield the original matrix) in ‘wplot’.

SRC_w2w/util_w2w.F: Correct the read-in

SRC_wplot/modules_rc.F: Remove the additional transpose

According to the Wien2k UG, the ROTLOC matrix R transforms r in the
global system to r' in the atom's local system, r' = R r.  In
wien2wannier, this affects the initial projections A_mn in ‘w2w’, and
the evaluation of the WFs in ‘wplot’.

To clarify wien2wannier's interpretation as of this commit, consider
an example.  Let

      / 0 1 0 \    / x → z
  R = | 0 0 1 | ~ {  y → x
      \ 1 0 0 /    \ z → y

and ‘d-xy’ be given as initial projection.  The initial projection
evaluated in the local system appears as

  d_xy(r') = d_xy(R r) ~ x' y' = z x ~ d_xz(r)

in the global system.  Thus, the resulting WF will tend to an xz-like
lobe.  This can be seen via ‘wplot’ or through the symmetries in
‘case_hr.dat’.